% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_cdisc_blocks.R
\name{new_random_adam_block}
\alias{new_random_adam_block}
\title{Random CDISC ADaM Block}
\usage{
new_random_adam_block(dataset = "qs", seed = 1, ...)
}
\arguments{
\item{dataset}{Dataset to generate (default "qs")}

\item{seed}{Random seed for reproducibility}

\item{...}{Additional arguments passed to new_block}
}
\description{
Generates a random ADaM dataset using random.cdisc.data package.
Requires ADSL as input.
}
