% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_cdisc_blocks.R
\name{new_random_adsl_block}
\alias{new_random_adsl_block}
\title{Random CDISC ADSL Block}
\usage{
new_random_adsl_block(seed = 1, n_subjects = 400, ...)
}
\arguments{
\item{seed}{Random seed for reproducibility}

\item{n_subjects}{Number of subjects (default 400)}

\item{...}{Additional arguments passed to new_block}
}
\description{
Generates a random ADSL dataset using random.cdisc.data package.
}
